requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-reviews==1.2.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-flex-fields==0.9.8
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
PyJWT==2.4.0
python-magic==0.4.15
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.1
